requirements.txt
dj-database-url==0.4.1
Django==1.9.8
djangorestframework==3.4.1
et-xmlfile==1.0.1
gunicorn==19.6.0
jdcal==1.2
openpyxl==2.3.5
psycopg2==2.6.2
whitenoise==2.0.6
